refactor: move Bytes classes to Bytes folder
closes #255
