+++ src/Yaapii.Atoms/Bytes/Base64Bytes.cs
@@ -1,23 +1,44 @@
-﻿using System;
-using System.Collections.Generic;
+﻿// MIT License
+//
+// Copyright(c) 2017 ICARUS Consulting GmbH
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+using System;
 using System.Text;
 
 namespace Yaapii.Atoms.Bytes
 {
     /// <summary>
-    /// Decodes all origin bytes using the Base64 encoding scheme.
+    /// Origin bytes decoded using the Base64 encoding scheme.
     /// </summary>
     public sealed class Base64Bytes : IBytes
     {
-        private readonly IBytes _bytes;
+        private readonly IBytes bytes;
 
         /// <summary>
-        /// Makes decoded origin bytes using the Base64 encoding scheme.
+        /// Origin bytes decoded using the Base64 encoding scheme.
         /// </summary>
         /// <param name="bytes">origin bytes</param>
         public Base64Bytes(IBytes bytes)
         {
-            _bytes = bytes;
+            this.bytes = bytes;
         }
 
         /// <summary>
@@ -26,7 +47,7 @@ namespace Yaapii.Atoms.Bytes
         /// <returns></returns>
         public byte[] AsBytes()
         {
-            var bytes = _bytes.AsBytes();
+            var bytes = this.bytes.AsBytes();
             string base64String = Encoding.UTF8.GetString(bytes, 0, bytes.Length);
             return Convert.FromBase64String(base64String);
         }
