+++ src/Yaapii.Atoms/Bytes/BytesEqual.cs
@@ -1,7 +1,24 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿// MIT License
+//
+// Copyright(c) 2017 ICARUS Consulting GmbH
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
 
 namespace Yaapii.Atoms.Bytes
 {
@@ -10,8 +27,8 @@ namespace Yaapii.Atoms.Bytes
     /// </summary>
     public sealed class BytesEqual : IScalar<bool>
     {
-        private readonly IBytes _left;
-        private readonly IBytes _right;
+        private readonly IBytes left;
+        private readonly IBytes right;
 
         /// <summary>
         /// Makes a truth about <see cref="IBytes"/> are equal or not.
@@ -20,8 +37,8 @@ namespace Yaapii.Atoms.Bytes
         /// <param name="right"></param>
         public BytesEqual(IBytes left, IBytes right)
         {
-            _left = left;
-            _right = right;
+            this.left = left;
+            this.right = right;
         }
 
         /// <summary>
@@ -30,8 +47,8 @@ namespace Yaapii.Atoms.Bytes
         /// <returns></returns>
         public bool Value()
         {
-            var left = _left.AsBytes();
-            var right = _right.AsBytes();
+            var left = this.left.AsBytes();
+            var right = this.right.AsBytes();
             var equal = left.Length == right.Length;
 
             for(var i=0;i<left.Length && equal;i++)
