+++ tests/Yaapii.Atoms.Tests/Bytes/BytesOfTest.cs
